refactor: 패키지명 대문자 제거
