+++ src/main/java/domain/board/strategy/MaSangSangMa.java
@@ -1,4 +1,4 @@
-package domain.board.createStrategy;
+package domain.board.strategy;
 
 import domain.Coordinate;
 import domain.Team;
@@ -9,7 +9,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class SangMaSangMa extends BoardCreateStrategy {
+public class MaSangSangMa extends BoardCreateStrategy {
 
     @Override
     public Map<Coordinate, Piece> createPiecesByTeam(Team team) {
@@ -17,8 +17,8 @@ public class SangMaSangMa extends BoardCreateStrategy {
         int y = TEAM_Y_COORDINATE.get(team);
 
         Map<Coordinate, Piece> pieces = new HashMap<>();
-        pieces.put(new Coordinate(x.get(0), y), new Sang(team));
-        pieces.put(new Coordinate(x.get(1), y), new Ma(team));
+        pieces.put(new Coordinate(x.get(0), y), new Ma(team));
+        pieces.put(new Coordinate(x.get(1), y), new Sang(team));
         pieces.put(new Coordinate(x.get(2), y), new Sang(team));
         pieces.put(new Coordinate(x.get(3), y), new Ma(team));
         return pieces;
